+++ src/yajl/yajl_tree.c
@@ -311,7 +311,7 @@ static int handle_number (void *ctx, const char *string, size_t string_length)
 
     errno = 0;
     v->u.number.i = yajl_parse_integer((const unsigned char *) v->u.number.r,
-                                       strlen(v->u.number.r));
+                                       (unsigned int)strlen(v->u.number.r));
     if (errno == 0)
         v->u.number.flags |= YAJL_NUMBER_INT_VALID;
 
@@ -456,8 +456,8 @@ yajl_val yajl_tree_get(yajl_val n, const char ** path, yajl_type type)
 {
     if (!path) return NULL;
     while (n && *path) {
-        unsigned int i;
-        int len;
+        size_t i;
+        size_t len;
 
         if (n->type != yajl_t_object) return NULL;
         len = n->u.object.len;
